update project structures
