+++ .gitignore
@@ -1,1 +1,9 @@
 .DS_Store
+.vscode
+build/
+CMakeFiles/
+
+cmake_install.cmake
+CMakeCache.txt
+Makefile
+tiny-ray-tracer
